Move constant variables and class
